requirements.txt
pyautogui
pillow
requests
pynput
